Developed first iteration of first question
