+++ src/logs/2024-05-01.txt
@@ -4,3 +4,7 @@ Log file created
 2024-05-01 21:47:34 - INFO - BigQuery client successfully initialized.
 2024-05-01 21:47:44 - INFO - BigQuery client successfully initialized.
 2024-05-01 21:48:27 - INFO - BigQuery client successfully initialized.
+2024-05-01 21:56:06 - INFO - BigQuery client successfully initialized.
+2024-05-01 21:57:31 - INFO - BigQuery client successfully initialized.
+2024-05-01 21:58:15 - INFO - BigQuery client successfully initialized.
+2024-05-01 21:58:17 - INFO - BigQuery client successfully initialized.
