+++ src/q1_time.py
@@ -1,5 +1,10 @@
 from typing import List, Tuple
 from datetime import datetime
+from db.bigquery import BigQueryClient
+from utils.queries import q1
 
 def q1_time(file_path: str) -> List[Tuple[datetime.date, str]]:
-    pass
+    bigquery_client = BigQueryClient()
+    result = bigquery_client.read_table(sql_query=q1)
+    list_output = [(row[1], row[0]) for row in result]
+    return list_output
